analysis/requirements.txt
yfinance
pandas
newspaper3k
lxml_html_clean
numpy
pandas
scikit-learn
spacy
yfinance
transformers
joblib
openpyxl
